Add support for mgmt cluster be managed
If management cluster is managed by another cluster we need to
differentiate which classifier instances sveltos-agent needs to
process and which classifierReports to collect.

When deploying Classifier istances, add annotation to indicate instance
was deployed by Sveltos and that sveltos-agent needs to process it.

Classifier instances created by platform/tenant admins won't have this
annotation.

If mgmt cluster is in turn managed by another cluster:
1. the classifier instances defined directly in the mgmt cluster (to be
evaluated in the managed cluster) won't have the annotation.
sveltos-agent running in the mgmt cluster will ignore those;
2. the classifier instances defined in the other cluster to be evaluated
in the management cluster will have this annotation (sveltos in the
other cluster will add it before deploying to mgmt cluster).
Sveltos-agent running in the management cluster will evaluate those.

ClassifierReports fecthed from managed cluster that have Spec Cluster
info set will be ignored (it is assumed those are already pulled from
another managed cluster as that info is only added when reports are
fecthed or pushed to mgmt cluster)