Start sveltos-agent in the management cluster
Classifier has new option: `agent-in-mgmt-cluster` when set classifier
will deploy svekltos-agent in the management cluster.

Resources deployed in the management cluster are:
- service
- deployment

All CRDs used by sveltos-agent and instances of those will still be
created in the managed cluster as the seltos-agent will connect to
the managed cluster.

When cluster is deleted, any sveltos-agent resource created in the
management cluster will be removed.