Sveltos-agent configuration
Sometimes it is useful to customize sveltos-agent deployment.
Classifier accepts a new arg: `sveltos-agent-config`

That is the name of the ConfigMap in the `projectsveltos` namespace
that contains patches that will be applied to sveltos-agent
before it gets deployed in the managed cluster.

For instance, creating this ConfigMap

```yaml
apiVersion: v1
data:
  patch: |-
    apiVersion: apps/v1
    kind: Deployment
    metadata:
      name: sveltos-agent
    spec:
      template:
        spec:
          containers:
          - name: manager
            image: projectsveltos/sveltos-agent:dev
            resources:
              requests:
                memory: 512Mi
            securityContext:
              readOnlyRootFilesystem: true
kind: ConfigMap
metadata:
  name: sveltos-agent
  namespace: projectsveltos
```

and setting the following extra argurment to classifier deployment

```
      - args:
        ...
        - --sveltos-agent-config=sveltos-agent
```

the sveltos-agent will be deployed in each managed cluster with
512Mi for request memory and image `projectsveltos/sveltos-agent:dev`