(bug) CRDs with agent less mode
When Sveltos is deployed in agentless mode (sveltos agent and drift
detection manager running in the management cluster), all Sveltos
resources used for events, health checks, configuration drift detection
and more, are created in the management cluster itself.
So in this mode there is no need to deploy Sveltos CRDs in each managed
cluster.