(bug) cleanup of stale resources
When agents operate from the management cluster, a dedicated sveltos-agent
deployment is created for each managed cluster.

A SveltosCluster and a Cluster controller are designed to automatically remove
these sveltos-agent deployments when a cluster is deleted. However, because
there are no Sveltos finalizers on the SveltosCluster or the ClusterAPI cluster
instances, it is possible for the classifier to miss a deletion event— for example,
if it's temporarily offline.

This update introduces a new logic to periodically scan for and remove any
sveltos-agent deployments that are "stale," meaning they correspond to clusters that
no longer exist. This ensures that agents are correctly cleaned up, even if a deletion
notification is missed.