(feat) Per-Cluster Patches for sveltos-agent and sveltos-applier
This update introduces the ability to apply unique configuration patches to
the sveltos-agent and sveltos-applier resources for individual managed clusters.

Sveltos deploys a sveltos agent (and a sveltos-applier when in pull mode) in every managed cluster
While patching the default configuration was supported, the same patch was applied uniformly across
all managed clusters.

This enhancement addresses scenarios requiring cluster-specific configuration overrides.
For example, a cluster located in a specific geographic region might need to pull the image
from a different private container registry than a cluster in another region.

Managed clusters can now define the `sveltosagent.projectsveltos.io/config-override-ref`
and `sveltosapplier.projectsveltos.io/config-override-ref` annotations to reference a
ConfigMap containing the desired patches.

The ConfigMap reference supports the following formats:

1. `<namespace>/<name>`: Explicitly specifies the namespace and name of the ConfigMap.
2. `<name> (or /<name>)`: Assumes the ConfigMap is in the same namespace as the annotated Cluster instance.

As with global override, the configuration data within the ConfigMap must be a patch of one of the
following types:
1. Strategic Merge Patch
2. JSON Patch (RFC6902)